ai/pyproject.toml
[project]
name = "deep-quest-ai"
version = "0.0.1"
description = "Deep Quest AI server"
authors = [{ name = "정현우", email = "smartcow.jung@gmail.com" }]

readme = "README.md"

requires-python = ">=3.11"

dependencies = [
    "langgraph-cli>=0.3.4",
    "langgraph>=1.0.4",
    "langgraph-api>=0.6.0",
    "python-dotenv>=1.1.1",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=3.2.0",
    "langchain-core>=1.1.0",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "langchain-openai>=1.1.0",
    "langgraph-sdk>=0.2.10",
    "langsmith>=0.4.48",
    "accelerate>=1.8.1",
    "aiofiles>=24.1.0",
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.14",
    "aiosignal>=1.4.0",
    "aiosqlite>=0.21.0",
    "annotated-types>=0.7.0",
    "antlr4-python3-runtime>=4.9.3",
    "anyio>=4.9.0",
    "attrs>=25.3.0",
    "backoff>=2.2.1",
    "blockbuster>=1.5.24",
    "cachetools>=5.5.2",
    "certifi>=2025.7.14",
    "cffi>=1.17.1",
    "chardet>=5.2.0",
    "charset-normalizer>=3.4.2",
    "click>=8.2.1",
    "cloudpickle>=3.1.1",
    "coloredlogs>=15.0.1",
    "contourpy>=1.3.2",
    "cryptography>=44.0.3",
    "cycler>=0.12.1",
    "dataclasses-json>=0.6.7",
    "deprecated>=1.2.18",
    "deprecation>=2.1.0",
    "distro>=1.9.0",
    "e>=1.4.5",
    "effdet>=0.4.1",
    "emoji>=2.14.1",
    "filelock>=3.18.0",
    "filetype>=1.2.0",
    "flatbuffers>=25.2.10",
    "fonttools>=4.58.5",
    "forbiddenfruit>=0.1.4",
    "frozenlist>=1.7.0",
    "fsspec>=2025.5.1",
    "google-ai-generativelanguage>=0.6.18",
    "google-api-core>=2.25.1",
    "google-auth>=2.40.3",
    "google-cloud-vision>=3.10.2",
    "googleapis-common-protos>=1.70.0",
    "grpcio>=1.73.1",
    "grpcio-status>=1.73.1",
    "h11>=0.16.0",
    "h2>=4.2.0",
    "hf-xet>=1.1.5",
    "hpack>=4.1.0",
    "html5lib>=1.1",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.1",
    "huggingface-hub>=0.33.2",
    "humanfriendly>=10.0",
    "hyperframe>=6.1.0",
    "idna>=3.10",
    "iniconfig>=2.1.0",
    "jinja2>=3.1.6",
    "jiter>=0.10.0",
    "joblib>=1.5.1",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "jsonschema-rs>=0.29.1",
    "kiwisolver>=1.4.8",
    "langchain>=1.1.0",
    "langchain-text-splitters>=0.3.8",
    "langdetect>=1.0.9",
    "langgraph-checkpoint>=2.1.0",
    "langgraph-prebuilt>=1.0.5",
    "langgraph-runtime-inmem>=0.5.0",
    "lxml>=6.0.0",
    "markupsafe>=3.0.2",
    "marshmallow>=3.26.1",
    "matplotlib>=3.10.3",
    "mpmath>=1.3.0",
    "multidict>=6.6.3",
    "mypy>=1.16.1",
    "mypy-extensions>=1.1.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.5",
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "olefile>=0.47",
    "omegaconf>=2.3.0",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.0",
    "openai>=1.97.0",
    "ormsgpack>=1.10.0",
    "packaging>=25.0",
    "pandas>=2.3.1",
    "pathspec>=0.12.1",
    "pi-heif>=1.0.0",
    "pillow>=11.3.0",
    "pip>=25.1.1",
    "pluggy>=1.6.0",
    "portalocker>=2.10.1",
    "propcache>=0.3.2",
    "proto-plus>=1.26.1",
    "protobuf>=6.31.1",
    "psutil>=7.0.0",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycocotools>=2.0.10",
    "pycparser>=2.22",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pydantic-settings>=2.10.1",
    "pygments>=2.19.2",
    "pyjwt>=2.10.1",
    "pyparsing>=3.2.3",
    "pytest>=8.4.1",
    "python-dateutil>=2.9.0.post0",
    "python-iso639>=2025.2.18",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "python-oxmsg>=0.0.2",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "rapidfuzz>=3.13.0",
    "regex>=2024.11.6",
    "redis>=6.4.0",
    "requests>=2.32.4",
    "requests-toolbelt>=1.0.0",
    "rsa>=4.9.1",
    "ruff>=0.12.2",
    "safetensors>=0.5.3",
    "scipy>=1.16.0",
    "setuptools>=80.9.0",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "soupsieve>=2.7",
    "sqlalchemy>=2.0.41",
    "sqlite-vec>=0.1.6",
    "sse-starlette>=2.1.3",
    "starlette>=0.47.1",
    "strenum>=0.4.15",
    "structlog>=25.4.0",
    "sympy>=1.14.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.9.0",
    "timm>=1.0.16",
    "tokenizers>=0.21.2",
    "tqdm>=4.67.1",
    "truststore>=0.10.1",
    "typing-extensions>=4.14.1",
    "typing-inspect>=0.9.0",
    "typing-inspection>=0.4.1",
    "tzdata>=2025.2",
    "urllib3>=2.5.0",
    "uvicorn>=0.35.0",
    "watchfiles>=1.1.0",
    "webencodings>=0.5.1",
    "websockets>=15.0.1",
    "wrapt>=1.17.2",
    "xxhash>=3.5.0",
    "yarl>=1.20.1",
    "zstandard>=0.23.0",
    "langchain-anthropic>=1.2.0",
    "google-genai>=1.31.0",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
"utils",
"common",
"common.schemas",
"resume_parser",
"jd_to_text", "jd_structuring", "question_gen", "question_feedback_gen"]

[tool.setuptools.package-dir]
"utils" = "src/utils"
"common" = "src/common"
"resume_parser" = "src/graphs/resume_parser"
"jd_to_text" = "src/graphs/jd_to_text"
"jd_structuring" = "src/graphs/jd_structuring"
"question_gen" = "src/graphs/question_gen"
"question_feedback_gen" = "src/graphs/question_feedback_gen"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
    # Disable mandatory docstrings
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "D106",  # Missing docstring in public nested class
    "D107",  # Missing docstring in __init__
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP", "T201"]  # Allow print statements in tests
"src/graphs/jd_to_text/*" = ["ALL"]  # Ignore all ruff checks for jd_to_text
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
exclude = [
    "src/graphs/jd_to_text/",  # Exclude jd_to_text from mypy checks
    "tests/integration_tests/",  # Exclude integration tests from mypy checks
]

[tool.pyrefly]
search_path = ["src", "src/graphs"]

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli>=0.4.10",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
    "types-requests>=2.32.4.20250913",
]
